src/test/java/_0801_0850/_819_Most_Common_Word/SolutionITest.java
package _0801_0850._819_Most_Common_Word;

import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

import static org.junit.jupiter.api.Assertions.assertEquals;

/**
 * @author jacka
 * @version 1.0 on 8/31/2020
 */
class SolutionITest {

  @ParameterizedTest
  @MethodSource("solutionStream")
  void testOnlineCase1(Solution solution) {
    final String paragraph = "Bob hit a ball, the hit BALL flew far after it was hit.";
    final String[] banned = {"hit"};
    assertEquals("ball", solution.mostCommonWord(paragraph, banned));
  }

  static Stream<Solution> solutionStream() {
    return Stream.of(
        new SolutionI()
    );
  }
}